+++ src/hooks/index.ts
@@ -1,6 +1,6 @@
 export * from './useBreakpoint/useBreakpoint';
 export * from './useIsMobile/useIsMobile';
-export * from './useIsomorphicLayoutEffect/useIsomorphicLayouEffect';
+export * from './useIsomorphicLayoutEffect/useIsomorphicLayoutEffect';
 export * from './useOpenClose/useOpenClose';
 export * from './useTheme/useTheme';
 export * from './useWindowSize/useWindowSize';
